move functions into new flowmapping.py file and into sectormapping.py, delete naics.py
